Added a simple test for the move queue
Added a test for the move queue. Adjusted some of the inout parameters
to account for the fact that they weren't actually being mutated, or
that those mutations weren't propagated back. Some of these could have
been addressed by changing the types to classes (so that they are
treated as pass by reference), but using structs is "more Swifty" - so
I'd prefer to keep them as such, and make other accommodations to allow
for that.