Added executeActions tests, refactored levels
Action levels are now set via a dictionary, which also dictates which
actions are enabled (if an action is found in the dictionary, it is
assumed to be enabled). Added tests for executeActions, though these
are very rudimentary at the moment and only test actions in isolation.
Testing a combination of actions is complicated due to the probabilities
involved, though these could be removed via dependency injection.