Removed C-style enum length check in favour of Swifty approach
Made a note about a possible issue with the renumbering of connections,
which wouldn't allow for arbitrarily disabled actions/senses. As long
as actions/senses are disabled starting from the end of the list it
should be fine. It'd be nice to support arbitrary disabling though - and
could be done via a mapping.