Added very basic controls, separated drawing and running
Added some simple controls mostly as a proof of concept. It is possible
to stop the simulation, and advance through it one step at a time. Also,
to reduce input lag, drawing of the simulation (which is currently
quite slow) has been separated from running the simulation (which is
comparatively fast).