Fixed a major bug in simulator steps
Mutating the contents of a value type (like an array or a struct) does
not change the original, so the previous implementation was not updating
the individuals stored in `peeps.individuals`. This wasn't noticed,
because some things (like movement) are handled by class types like Peep
or by static instances like Grid.