Reduced usage of inout
Reduced the usage of inout, in favour of explicitly returning new
values. This could have some performance implications, but has the
advantage of making it more obvious what has changed and what needs
to be persisted to keep that change. Some more complex inouts are
remaining, and some were superfluous.