Greatly increased performance by changing signal fading
Signal fading was a bottleneck in performance due to the large amount of
values that needed to be updated (each element in the grid). Leveraged a
Single Instruction Multiple Data (SIMD) library to reduce the time to
less than 20% of what it originally took. Unfortunately, this library
doesn't provide any obvious way of clipping the signals, so instead of
clipping them to 0, I've changed the values to floating point and apply
damping (a multiplier less than 1.0) to fade them. This differs from
the original implementation, but should be sufficient.