Challenges have been converted to separate types
Challenges are now implemented as types conforming to a protocol rather
than an enum. This helps readability by avoiding a very large enum,
but it also allows for easier addition of new challenges without
impacting others, and allows for challenges to be parameterized (see
CornerChallenge for an example, which allows both weighted and
unweighted scoring based on parameters). This would also allow for
composition of challenges (ex. RightQuarter & LeftQuarter). End of sim
step has also been changed, and challenges are now applied as modifiers
to an individuals actions - this may be slightly more performant by
avoiding a separate loop.