Updated tests to account for changes to ExecuteActions
Removed some tests that cover logic already covered elsewhere, this
is one of the benefits of having reusable actions. Also, some edge
cases will be tested in files more targeted towards specific actions.
Currently having an issue with the KillAction which is touching
global state - will need to investigate that.