Removed Peeps, Grid now uses a dictionary.
Removed, or more accurately, combined Peeps with the Grid. So the
individuals no longer exist separately from their accompanying grid.
The grid now uses a dictionary to keep track of which individuals are
where, which has pros and cons but I think the pros outweigh the cons.
Fixed KillAction and it's related test by no longer using global state.