Only update nodes after finished advancing steps
Rather than trying to redraw the nodes every frame, it's more performant
to draw them only after all the steps have been processed. For the
case where the steps per refresh are 1, this is the same as before
(update notes every frame), but for any higher values, it results in
a number of updates being skipped. This can make it look a choppier,
but the number of steps being processed is greater (and the stats label
is still updated in realtime).