Reverted removal of setting current location
I forgot that some things (like the challenge bits) were being updated
by overwriting the individual at the previous current location with the
values from the updated individual. So this can't be removed. Should
make some changes such that it's not possible for something to change
the location of the individual without using queueForMove though.