Fixed an issue where cells could "freeze"
Only updating living cells comes with the problem that any other cells
"freeze" at whatever their last state was, leading to "ghosts". This
could also be mitigated by updating both the living and dead cells, but
that would result in the same problem if it was possible for the
population to go up at runtime. Instead, we iterate over all cells
and hide them if they're not tied to a living individual.