Implemented a quick way to view history
All steps for the current generation are kept, while only the last
steps of each prior generation is kept to conserve memory. Rewinding
cannot be undone except by allowing the simulation to play out again,
but that might yield different results due to randomness. Ultimately
while this works, I think saving (and compressing) textual data might
be a nicer long term solution as it would allow for all of the
intermediate steps to be kept. Either that, or a stripped down version
of the individuals (without their neurons etc.) could be saved.