Encapsulated grid, signals and most mutable state into a "World" type
Most mutable state is now encapsulated into a "World" type. Rather
than relying on global state, other types now rely on being passed a
World, which makes it easier to ensure that (for example) when rewinding
time, everything is put back to the state it was in at that time
(previously, signals for example weren't being reset). Introduced a
simulator type which helps to advance from one world to another, with
each world representing a snapshot of time. Each world also carries it's
parameters with it, allowing for parameters to be seamlessly mutated
over time and restored with the rewind feature.