History can now be traversed in both directions
One can go forward and backward in history now, which is helpful for
debugging. Rather than storing separate long/short term histories,
the history is just clipped to a specific size (or unlimited, if
desired). This was simpler to implement, though a rough approximation
of the previous solution could be implemented by only saving every nth
step.