Improved consistency of tests
Some tests were verifying random outcomes, which isn't ideal. In some
cases this resulted in flaky tests, and in other cases it was addressed
with looping to eventually find the result we wanted. Improved this
by exposing the probability curves as a parameter, such that it could
be stubbed in tests. Doesn't cover all randomness, but covers what is
currently tested.