Tests using genome are now consistent
Swift's Dictionary type (as defined in the standard library) does not
guarantee ordering of it's keys, and so the node map's elements were
being modified in a non-deterministic manner, making it difficult to
test as the results could vary wildly depending on how the nodes are
numbered. Using a library to provide an OrderedDictionary type would
make this more explicit, and possibly improve performance, but this
works for now.