Initial attempt at sharing state between SwiftUI and SpriteKit
Some state is now configured in SwiftUI, and accessible to SpriteKit.
This does not currently work in reverse however, so SwiftUI doesn't
become aware of any changes to the state made by SpriteKit. This can
be addressed by adding @Published property wrappers, but that results in
the scene being recreated. Need to investigate this further.