Introduced a custom Size type
Introduced a Size type instead of using a tuple, allowing for automatic
Equatable conformance of types that make use of this one. Changed
property names to be more descriptive (width vs x). This type only
really differs from CGSize in that it is intended for use with integer
values (whereas CGSize deals with floating point). A name like
`IntegralSize` might be more suitable if this leads to confusion.